backend2/requirements.txt
fastapi==0.70.1
uvicorn==0.22.0
python-keycloak==0.24.0
requests==2.30.0
urllib3==1.26.7
python-multipart==0.0.5
simber==0.2.1
SQLAlchemy==1.4.14
pydantic==1.8.2
psycopg2-binary==2.8.6
alembic==1.6.2
faker==8.1.2
# fastapi_keycloak==1.0.0
PyJWT==2.8.0
